requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
chromadb
pypdf
python-dotenv
pyyaml
tiktoken
pandas
streamlit
langchain-huggingface
sentence-transformers
